Split MCP server with traits
